\name{parameters_col_metoxi}
\alias{parameters_col_metoxi}
\docType{data}
\title{
 Parameters data frame for sulphadimetoxine columnar measurements.
}
\description{
 Data frame containing the parameters calculated with processPeak function for columnar measurements of sulphadimetoxine.
}
\usage{data("parameters_col_metoxi")}
\format{
  A data frame with 12 observations on the following 9 variables.
  \describe{
    \item{\code{compound}}{a factor with levels \code{metoxi}}
    \item{\code{flow}}{a numeric vector}
    \item{\code{tr}}{a numeric vector}
    \item{\code{Hmax}}{a numeric vector}
    \item{\code{A60}}{a numeric vector}
    \item{\code{B60}}{a numeric vector}
    \item{\code{A10}}{a numeric vector}
    \item{\code{B10}}{a numeric vector}
    \item{\code{area}}{a numeric vector}
  }
}
\keyword{datasets}
