\name{parameters_dead}
\alias{parameters_dead}
\docType{data}
\title{
 Parameters data frame for dead marker measurements.
}
\description{
 Data frame containing the parameters calculated with processPeak function for dead marker measurements.
}
\usage{data("parameters_dead")}
\format{
  A data frame with 13 observations on the following 9 variables.
  \describe{
    \item{\code{compound}}{a factor with levels \code{Kbr}}
    \item{\code{flow}}{a numeric vector}
    \item{\code{tr}}{a numeric vector}
    \item{\code{Hmax}}{a numeric vector}
    \item{\code{A60}}{a numeric vector}
    \item{\code{B60}}{a numeric vector}
    \item{\code{A10}}{a numeric vector}
    \item{\code{B10}}{a numeric vector}
    \item{\code{area}}{a numeric vector}
  }
}
\keyword{datasets}
