\name{parameters_ext}
\alias{parameters_ext}
\docType{data}
\title{
 Parameters data frame for Kbr extracolumnar measurements.
}
\description{
  Data frame containing the parameters calculated with processPeak function for extracolumnar measurements of kbr.
}
\usage{data("parameters_ext")}
\format{
  A data frame with 13 observations on the following 9 variables.
  \describe{
    \item{\code{compound}}{a factor with levels \code{Kbre}}
    \item{\code{flow}}{a numeric vector}
    \item{\code{tr}}{a numeric vector}
    \item{\code{Hmax}}{a numeric vector}
    \item{\code{A60}}{a numeric vector}
    \item{\code{B60}}{a numeric vector}
    \item{\code{A10}}{a numeric vector}
    \item{\code{B10}}{a numeric vector}
    \item{\code{area}}{a numeric vector}
  }
}
\keyword{datasets}
