\name{peak}
\alias{peak}
\docType{data}
\title{
 Peak read using readChrom function.
}
\description{
 Data frame with 2 columns (x, y) corresponding to a chromatographic peak that has been read and extracted using readChrom function.
}
\usage{data("peak")}
\format{
  A data frame with 12000 observations on the following 2 variables.
  \describe{
    \item{\code{V1}}{a numeric vector}
    \item{\code{V2}}{a numeric vector}
  }
}
\keyword{datasets}
